chore: refactor namespaces
